Update adapter with a Read file. Added copyright to all files
